# api/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "HTA2 GANTT APIs",
        "title": "HTA GANTT APIs",
        "contact": {},
        "license": {
            "name": "AGPL 3.0",
            "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
        },
        "version": "0.1"
    },
    "host": "pmip.t.api.likbox.com",
    "basePath": "/hta-gantt/v1.0",
    "paths": {
...
        "/resources": {
            "post": {
                "description": "新增資源",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resource"
                ],
                "summary": "新增資源",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWE Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "新增資源",
                        "name": "*",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/resources.Create"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功後返回的值",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.SuccessfulMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "415": {
                        "description": "必要欄位帶入錯誤",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.ErrorMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "detailed": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "伺服器非預期錯誤",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.ErrorMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "detailed": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/resources/import": {
            "post": {
                "description": "匯入資源",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resource"
                ],
                "summary": "匯入資源",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWE Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "匯入資源",
                        "name": "*",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/resources.Import"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功後返回的值",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.SuccessfulMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "415": {
                        "description": "必要欄位帶入錯誤",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.ErrorMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "detailed": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "伺服器非預期錯誤",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.ErrorMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "detailed": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/resources/list": {
            "post": {
                "description": "取得全部資源",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resource"
                ],
                "summary": "取得全部資源",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWE Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "目前頁數,請從1開始帶入",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "一次回傳比數,請從1開始帶入,最高上限20",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "搜尋",
                        "name": "*",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/resources.Filter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功後返回的值",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.SuccessfulMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/resources.List"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "415": {
                        "description": "必要欄位帶入錯誤",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.ErrorMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "detailed": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "伺服器非預期錯誤",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.ErrorMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "detailed": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/resources/no-pagination": {
            "get": {
                "description": "取得全部資源",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resource"
                ],
                "summary": "取得全部資源(不用page\u0026limit)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWE Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功後返回的值",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.SuccessfulMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/resources.List"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "415": {
                        "description": "必要欄位帶入錯誤",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.ErrorMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "detailed": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "伺服器非預期錯誤",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.ErrorMessage"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "detailed": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "code.ErrorMessage": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "回傳代碼",
                    "type": "integer"
                },
                "detailed": {
                    "description": "詳細錯誤內容"
                },
                "message": {
                    "description": "錯誤回傳訊息",
                    "type": "string"
                },
                "timestamp": {
                    "description": "錯誤時間",
                    "type": "string",
                    "example": "2021-07-29T07:23:47Z"
                }
            }
        },
        "code.SuccessfulMessage": {
            "type": "object",
            "properties": {
                "body": {
                    "description": "正確回傳內容"
                },
                "code": {
                    "description": "回傳代碼",
                    "type": "integer"
                },
                "timestamp": {
                    "description": "錯誤時間",
                    "type": "string",
                    "example": "2021-07-29T07:23:47Z"
                }
            }
        },
...
        "resources.Create": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "信箱",
                    "type": "string"
                },
                "is_expand": {
                    "type": "boolean"
                },
                "phone": {
                    "description": "電話",
                    "type": "string"
                },
                "resource_group": {
                    "description": "群組",
                    "type": "string"
                },
                "resource_name": {
                    "description": "名字",
                    "type": "string"
                },
                "standard_cost": {
                    "type": "number",
                    "minimum": 0
                },
                "tags": {
                    "description": "標籤",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "total_cost": {
                    "type": "number",
                    "minimum": 0
                },
                "total_load": {
                    "description": "總負載",
                    "type": "number",
                    "minimum": 0
                }
            }
        },
        "resources.Filter": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "信箱",
                    "type": "string"
                },
                "phone": {
                    "description": "電話",
                    "type": "string"
                },
                "resource_group": {
                    "description": "群組",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "resource_name": {
                    "description": "名字",
                    "type": "string"
                }
            }
        },
        "resources.Import": {
            "type": "object",
            "properties": {
                "base64": {
                    "description": "Base64",
                    "type": "string"
                }
            }
        },
        "resources.List": {
            "type": "object",
            "required": [
                "limit",
                "page"
            ],
            "properties": {
                "limit": {
                    "description": "筆數(請從1開始帶入,最高上限20)",
                    "type": "integer"
                },
                "page": {
                    "description": "頁數(請從1開始帶入)",
                    "type": "integer"
                },
                "pages": {
                    "description": "總頁數",
                    "type": "integer"
                },
                "resources": {
                    "description": "多筆",
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "created_at": {
                                "description": "創建時間",
                                "type": "string"
                            },
                            "created_by": {
                                "description": "創建者",
                                "type": "string"
                            },
                            "deleted_at": {
                                "description": "刪除時間",
                                "type": "string"
                            },
                            "email": {
                                "description": "信箱",
                                "type": "string"
                            },
                            "is_expand": {
                                "type": "boolean"
                            },
                            "phone": {
                                "description": "電話",
                                "type": "string"
                            },
                            "resource_group": {
                                "description": "v",
                                "type": "string"
                            },
                            "resource_id": {
                                "description": "舊編號 ＆ 頁面ID",
                                "type": "integer"
                            },
                            "resource_name": {
                                "description": "名字",
                                "type": "string"
                            },
                            "resource_uuid": {
                                "description": "表ID",
                                "type": "string"
                            },
                            "standard_cost": {
                                "type": "number"
                            },
                            "tags": {
                                "description": "標籤",
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "total_cost": {
                                "type": "number"
                            },
                            "total_load": {
                                "description": "總負載",
                                "type": "number"
                            },
                            "updated_at": {
                                "description": "更新時間",
                                "type": "string"
                            },
                            "updated_by": {
                                "description": "更新者",
                                "type": "string"
                            }
                        }
                    }
                },
                "total": {
                    "description": "總筆數",
                    "type": "integer"
                }
            }
        },
...
    }
}